cronDeliveryAPI/requirements.txt
asgiref             ==3.2.7
certifi             ==2020.4.5.1
cffi                ==1.14.0
chardet             ==3.0.4
cryptography        ==2.9.2
fcm-django          ==0.3.4
dj-database-url     ==0.5.0
Django              ==3.0.8
django-extensions   ==2.2.9
django-filter       ==2.2.0
django-nested-admin ==3.3.2
django-rest-knox    ==4.1.0
djangorestframework ==3.11.0
gunicorn            ==20.0.4
idna                ==2.9
Pillow              ==7.1.1
pip                 ==20.1
psycopg2            ==2.8.5
pycparser           ==2.20
PyJWT               ==1.7.1
pyotp               ==2.3.0
python-decouple     ==3.3
pytz                ==2019.3
requests            ==2.23.0
setuptools          ==41.2.0
six                 ==1.14.0
sqlparse            ==0.3.1
urllib3             ==1.25.9
whitenoise          ==5.0.1